+++ pulse-sys/src/format.rs
@@ -30,6 +30,10 @@ pub enum pa_encoding_t {
     MPEG_IEC61937,
     DTS_IEC61937,
     MPEG2_AAC_IEC61937,
+    #[cfg(any(feature = "pa_v13", feature = "dox"))]
+    TRUEHD_IEC61937,
+    #[cfg(any(feature = "pa_v13", feature = "dox"))]
+    DTSHD_IEC61937,
 
     Invalid = -1,
 }
@@ -43,6 +47,10 @@ pub const PA_ENCODING_EAC3_IEC61937: pa_encoding_t = pa_encoding_t::EAC3_IEC6193
 pub const PA_ENCODING_MPEG_IEC61937: pa_encoding_t = pa_encoding_t::MPEG_IEC61937;
 pub const PA_ENCODING_DTS_IEC61937: pa_encoding_t = pa_encoding_t::DTS_IEC61937;
 pub const PA_ENCODING_MPEG2_AAC_IEC61937: pa_encoding_t = pa_encoding_t::MPEG2_AAC_IEC61937;
+#[cfg(any(feature = "pa_v13", feature = "dox"))]
+pub const PA_ENCODING_TRUEHD_IEC61937: pa_encoding_t = pa_encoding_t::TRUEHD_IEC61937;
+#[cfg(any(feature = "pa_v13", feature = "dox"))]
+pub const PA_ENCODING_DTSHD_IEC61937: pa_encoding_t = pa_encoding_t::DTSHD_IEC61937;
 pub const PA_ENCODING_INVALID: pa_encoding_t = pa_encoding_t::Invalid;
 
 impl Default for pa_encoding_t {
@@ -114,6 +122,14 @@ extern "C" {
     pub fn pa_format_info_get_prop_string(f: *const pa_format_info, key: *const c_char, v: *mut *mut c_char) -> i32;
     pub fn pa_format_info_get_prop_string_array(f: *const pa_format_info, key: *const c_char, values: *mut *mut *mut c_char, n_values: *mut i32) -> i32;
     pub fn pa_format_info_free_string_array(values: *mut *mut c_char, n_values: i32);
+    #[cfg(any(feature = "pa_v13", feature = "dox"))]
+    pub fn pa_format_info_get_sample_format(f: *const pa_format_info, sf: *mut pa_sample_format_t) -> i32;
+    #[cfg(any(feature = "pa_v13", feature = "dox"))]
+    pub fn pa_format_info_get_rate(f: *const pa_format_info, rate: *mut u32) -> i32;
+    #[cfg(any(feature = "pa_v13", feature = "dox"))]
+    pub fn pa_format_info_get_channels(f: *const pa_format_info, channels: *mut u8) -> i32;
+    #[cfg(any(feature = "pa_v13", feature = "dox"))]
+    pub fn pa_format_info_get_channel_map(f: *const pa_format_info, map: *const pa_channel_map) -> i32;
     pub fn pa_format_info_set_prop_int(f: *mut pa_format_info, key: *const c_char, value: i32);
     pub fn pa_format_info_set_prop_int_array(f: *mut pa_format_info, key: *const c_char, values: *const i32, n_values: i32);
     pub fn pa_format_info_set_prop_int_range(f: *mut pa_format_info, key: *const c_char, min: i32, max: i32);